authorization implemented

+++ src/api/notes/handler.js
@@ -12,13 +12,14 @@ class NotesHandler {
 
   async postNoteHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
-    const { title = "untitled", body, tags } = request.payload;
+    const { title = 'untitled', body, tags } = request.payload;
+    const { id: owner } = request.auth.credentials;
 
-    const noteId = await this._service.addNote({ title, body, tags });
+    const noteId = await this._service.addNote({ title, body, tags, owner: credentialId });
 
     const response = h.response({
-      status: "success",
-      message: "Catatan berhasil ditambahkan",
+      status: 'success',
+      message: 'Catatan berhasil ditambahkan',
       data: {
         noteId,
       },
@@ -27,10 +28,11 @@ class NotesHandler {
     return response;
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: owner } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
-      status: "success",
+      status: 'success',
       data: {
         notes,
       },
@@ -39,9 +41,12 @@ class NotesHandler {
 
   async getNoteByIdHandler(request, h) {
     const { id } = request.params;
+    const { id: owner } = request.auth.credentials;
+
+    await this._service.verifyNoteOwner(id, credentialId);
     const note = await this._service.getNoteById(id);
     return {
-      status: "success",
+      status: 'success',
       data: {
         note,
       },
@@ -50,23 +55,27 @@ class NotesHandler {
 
   async putNoteByIdHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
+    const { id: owner } = request.auth.credentials;
     const { id } = request.params;
 
-    await this._service.editNoteById(id, request.payload);
+    await this._service.verifyNoteOwner(id, credentialId);
+    this._service.editNoteById(id, request.payload);
 
     return {
-      status: "success",
-      message: "Catatan berhasil diperbarui",
+      status: 'success',
+      message: 'Catatan berhasil diperbarui',
     };
   }
 
   async deleteNoteByIdHandler(request, h) {
     const { id } = request.params;
-    await this._service.deleteNoteById(id);
+    const { id: owner } = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, credentialId);
+    this._service.deleteNoteById(id);
 
     return {
-      status: "success",
-      message: "Catatan berhasil dihapus",
+      status: 'success',
+      message: 'Catatan berhasil dihapus',
     };
   }
 }
